+++ src/task/Subtask.java
@@ -15,21 +15,21 @@ public class Subtask extends Task {
     }
 
     public Subtask(
-            int uid, TaskType taskType, String name, Status status, int duration, LocalDateTime startTime,
+            int uid, TaskType taskType, String name, Status status, long duration, LocalDateTime startTime,
             String description, int uidOfEpic
     ) {
         super(uid, taskType, name, status, duration, startTime, description);
         this.uidOfEpic = uidOfEpic;
     } //Конструктор для загрузки сохранки
 
-    public Subtask(String name, String description, int uidOfEpic, String status, int duration, LocalDateTime startTime) {
+    public Subtask(String name, String description, int uidOfEpic, String status, long duration, LocalDateTime startTime) {
         super(name, description, TaskType.SUBTASK);
         this.uidOfEpic = uidOfEpic;
         this.duration = duration;
         this.startTime = startTime;
     }
 
-    public Subtask(String name, String description, int uidOfEpic, int duration, LocalDateTime startTime) {
+    public Subtask(String name, String description, int uidOfEpic, long duration, LocalDateTime startTime) {
         super(name, description, TaskType.SUBTASK);
         this.uidOfEpic = uidOfEpic;
         this.duration = duration;
